[impl] ipcam update. need to npm update // install.

+++ src/library/rtsp-ffmpeg/lib/rtsp-ffmpeg.js
@@ -7,6 +7,8 @@ const spawn = require('child_process').spawn
     , util = require('util')
 ;
 
+var moment = require('moment');
+
 /**
  * Stream constructor
  * @param {object} options
@@ -46,11 +48,16 @@ function newListener(event) {
     if (event === 'data' && this.listeners(event).length === 0) {
         this.start();
     }
+
+    // if (event === 'record' && this.listener(event).length() === 0 ){
+    //     this.record();
+    // }
 }
 
 function removeListener(event) {
     if (event === 'data' && this.listeners(event).length === 0) {
         this.stop();
+        console.log('remove Listener');
     }
 }
 
@@ -62,6 +69,7 @@ function observer(changes) {
     }
 }
 
+//
 // FFMpeg.prototype._args = function() {
 //     //영상 출력
 //     //ffmpeg -i rtsp:/211.238.240.86:554/unicast -r 10 -q:v 10 -f image2 -udatefirst 1 -
@@ -87,75 +95,110 @@ function observer(changes) {
 //     ]);
 // };
 
-var on_off_record = false;
-var path = require('path');
-const campath = path.join(__dirname, '../../../resource/public');
-
 FFMpeg.prototype._args = function (option) {
     //영상 출력 // 썸네일
-    console.log("yes");
-    var date = new Date().toTimeString().replace(/.*(\d{2}-\d{2}-\d{2}).*/, "$1");
+    var now = moment(Date.now()).format('YYYYMMDDHHmmss');
+    var recordFilePath = './resource/public/video/' + now + '.mp4';
+    var thumbnailPath = './resource/public/image/thumbnail.png';
 
-    var thumbnail = [
-        '-loglevel', 'quiet', '-y'
-        , '-i', this.input,
-        '-ss', '0',
-        '-vframes', '1', campath+'/image/thumbnail.png'
-    ];
-
-
-    var video1 = [
-        '-loglevel', 'quiet'
-        , '-i', this.input
-        , '-r', this.rate.toString()];
-
-    var video2 = this.quality ? ['-q:v', this.quality.toString()] : [];
-
-    var video3 = [
-        '-vf', 'scale=' + this.resolution,
-        //  '-b:v', '512k',
-        '-f', 'image2'
-        , '-updatefirst', '1'
-        , '-'
-    ];
-
-    var record = [
-        '-loglevel', 'quiet'
-        , '-i', this.input,
-        '-vcodec', 'copy',
-        '-ss', '0', '-t', '10',
-        campath + '/video/' + date +'.avi'
-    ];
-
-    var stream = this.arguments.concat(thumbnail, video1, video2, video3);
-    var record = this.arguments.concat(record);
-
-    var data;
-
-    if (option == 0) {
-        data = stream;
-        on_off_record = false;
-    } else if (option == 1) {
-        data = record;
-        on_off_record = true;
+    if (option === 0) {
+        return this.arguments.concat([
+            '-loglevel', 'quiet', '-y'
+            , '-i', this.input,
+            '-ss', '0',
+            '-vframes', '1', thumbnailPath
+        ]);
+    }
+    else if (option === 1) {
+        return this.arguments.concat([
+                '-loglevel', 'quiet', '-y'
+                , '-i', this.input,
+                '-ss', '0',
+                '-vframes', '1', thumbnailPath
+            ],
+            [
+                '-loglevel', 'quiet'
+                , '-i', this.input
+                , '-r', this.rate.toString()],
+            this.quality ? ['-q:v', this.quality.toString()] : [],
+            [
+                '-vf', 'scale=' + this.resolution,
+                //  '-b:v', '512k',
+                '-f', 'image2'
+                , '-updatefirst', '1'
+                , '-'
+            ]);
+    }
+    else if (option === 2) {
+        return this.arguments.concat(
+            ['-loglevel', 'quiet', '-y', '-i',
+                this.input, '-vcodec', 'copy', '-ss', '0',
+                '-t', '10', recordFilePath]);
     }
 
-    return data;
-}
+    // var thumbnail = [
+    //     '-loglevel', 'quiet', '-y'
+    //     , '-i', this.input,
+    //     '-ss', '0',
+    //     '-vframes', '1', thumbnailPath
+    // ];
+    //
+    //
+    // var video1 = [
+    //     '-loglevel', 'quiet'
+    //     , '-i', this.input
+    //     , '-r', this.rate.toString()];
+    //
+    // var video2 = this.quality ? ['-q:v', this.quality.toString()] : [];
+    //
+    // var video3 = [
+    //     '-vf', 'scale=' + this.resolution,
+    //     //  '-b:v', '512k',
+    //     '-f', 'image2'
+    //     , '-updatefirst', '1'
+    //     , '-'
+    // ];
+    //
+    // var stream = this.arguments.concat(thumbnail, video1, video2, video3);
+    // var record = this.arguments.concat(['-loglevel', 'debug', '-y', '-i',
+    //     this.input, '-vcodec', 'copy', '-ss', '0', '-t', '10', recordFilePath]);
+    //
+    // var data;
+    //
+    // if (option == 0) {
+    //     data = stream;
+    //     on_off_record = false;
+    // } else if (option == 1) {
+    //     data = record;
+    //     on_off_record = true;
+    // }
+    // return data;
+};
 
-FFMpeg.prototype.start = function () {
-
+FFMpeg.prototype.init = function() {
     this.child = spawn(FFMpeg.cmd, this._args(parseInt(0)));
     this.child.stdout.on('data', this.emit.bind(this, 'data'));
     this.child.stderr.on('data', function (data) {
-        throw new Error(data);
+        // console.log('error');
+        // console.log(data);
+        // throw new Error(data);
+    });
+};
+
+FFMpeg.prototype.start = function () {
+    this.child = spawn(FFMpeg.cmd, this._args(parseInt(1)));
+    this.child.stdout.on('data', this.emit.bind(this, 'data'));
+    this.child.stderr.on('data', function (data) {
+        // console.log('error ');
+        // throw new Error(data);
     });
 
-    this.on('record', function (record_num) {
+    // this.on('record', function (record_num) {
+    //
+    //     this.child = spawn(FFMpeg.cmd, this._args(parseInt(record_num)));
+    //     console.log("녹화시작?" + record_num);
+    // });
 
-        this.child = spawn(FFMpeg.cmd, this._args(parseInt(record_num)));
-        console.log("녹화시작?" + record_num);
-    })
     this.child.on('close', function (code) {
         if (code === 0) {
             setTimeout(FFMpeg.prototype.start.bind(this), 1000);
@@ -168,9 +211,11 @@ FFMpeg.prototype.start = function () {
  * Stop ffmpeg spawn process
  */
 FFMpeg.prototype.stop = function () {
-    this.child.kill();
-    delete this.child;
-    this.emit('stop');
+    if (this.child) {
+        this.child.kill();
+        delete this.child;
+        this.emit('stop');
+    }
 };
 /**
  * Restart ffmpeg spawn process
@@ -180,6 +225,33 @@ FFMpeg.prototype.restart = function () {
         this.stop();
         this.start(0);
     }
+};
+
+FFMpeg.prototype.record = function () {
+    this.child = spawn(FFMpeg.cmd, this._args(parseInt(2)));
+    // this.child.stdout.on('data', (data) => {
+    //     console.log(`stdout: ${data}`);
+    // });
+
+    this.child.stdout.on('data', function (data) {
+        console.log('success');
+        // console.log(data.toString());
+        this.emit.bind(data, 'data');
+    });
+    // this.child.stdout.on('data', this.emit.bind(this, 'data'));
+    this.child.stderr.on('data', function (data) {
+        console.log('error');
+        console.log(data.toString());
+        // throw new Error(data);
+    });
+
+    this.child.on('close', function (code) {
+        if (code === 0) {
+            setTimeout(FFMpeg.prototype.start.bind(this), 1000);
+        }
+    }.bind(this));
+
+    this.emit('start');
 };
 
 if (typeof Proxy === 'function') {
